Reduce SURB size
By replacing the first payload encryption key with a 32-byte secret from
which the key is derived.

It might be possible to derive the first payload encryption key from the
public key in the SURB header and then perform a blinding step to
produce the real public key. This would save another 32 bytes. I'm not
sure if this would be safe though, and it would require more elliptic
curve operations, which are very expensive.