Fix replay filter
Given any X25519 public key, it is trivial to construct keys that _look_
different but are actually equivalent:

- Add or subtract 2^255 - 19.
- Add a low order point.

Such keys could be detected, but it is simpler to just directly prevent
replay of shared secrets, so that is what I've done.

It's unclear to me whether or not it's safe to use SipHash to hash the
shared secrets, so switch to BLAKE2b. This is more expensive, but it
should make little difference overall.