Drop check of shared secret against identity
The Sphinx paper doesn't really say to do this; it says to check that
the public key is in the prime-order group (and not the identity
element), which is a stricter and much more expensive check.

X25519 is explicitly designed so that public keys need not be validated,
unless the protocol requires both parties to "contribute" to the shared
secret. I do not believe that is the case here, so simply drop the
validation.